move stop to separate file
